backend/html-report-service/requirements.txt
Flask
gunicorn
google-cloud-storage
Flask-Cors